flask/requirements.txt
flask
flask-kafka
flask_cors
kafka-python
Flask-SocketIO
eventlet
flask-sock